requirements.txt
pyodbc==4.0.26
setuptools==40.8.0
smartsheet_python_sdk==2.68.3
pandas==1.0.2
numpy==1.16.4
SQLAlchemy==1.3.4
protobuf==3.9.1
google_api_python_client==1.7.10
xlrd==1.2.0
azure-devops==6.0.0b2
sqlalchemy-redshift==0.7.7
retrying==1.3.3
boto3==1.13.16